requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
pefile==2024.8.26
python-magic==0.4.27
requests==2.32.5
urllib3==2.5.0
